properly skip vbr file tags
the decoder only skips tags at the beginning on the initial call to NewDecoder. So in order to properly handle a single stream of mp3s stiched together I need to skip the tags myself.

This can probably be refactored to use a buffered reader to properly unread the 3 bytes that always get checked. Right now this just removes the first 3 bytes of every file regardless. Not sure if that is really an issue or not.